move control to headless. #17
