+++ src/interactive_pipe/__init__.py
@@ -1,3 +1,3 @@
 from interactive_pipe.core.decorator import interactive, pipeline, interactive_pipeline
-from interactive_pipe.core.control import Control
+from interactive_pipe.headless.control import Control
 from interactive_pipe.headless.keyboard import  KeyboardControl
