+++ src/interactive_pipe/graphical/nb_control.py
@@ -1,5 +1,5 @@
 from ipywidgets import Dropdown, FloatSlider, IntSlider, Checkbox, Layout
-from interactive_pipe.core.control import Control
+from interactive_pipe.headless.control import Control
 import  logging
